criando autenticação de usuários

+++ src/routers/router_projetos.py
@@ -3,14 +3,14 @@ from schemas.schema import ProjetosResponse, ProjetosRequest
 from typing import List
 from sqlalchemy.orm import Session
 from connection.dependences import get_db
-from models.models import Projetos
-
+from models.models import Projetos, User
+from services.security import (get_current_user)
 
 router = APIRouter(prefix="/projetos")
 
 #retorna todos os projetos em forma de lista
 @router.get("/", response_model=List[ProjetosResponse],status_code=200)
-def get_projetos(db:Session = Depends(get_db)) -> ProjetosResponse:
+def get_projetos(db:Session = Depends(get_db),current_user: User = Depends(get_current_user)) -> ProjetosResponse:
     projetos = db.query(Projetos).all()
     if not projetos:
         raise HTTPException(status_code=404, detail="Não há projetos cadastrados")
@@ -19,7 +19,7 @@ def get_projetos(db:Session = Depends(get_db)) -> ProjetosResponse:
 
 
 @router.post("/", response_model=ProjetosResponse, status_code=201)
-def post_projetos(projetos: ProjetosRequest, db: Session = Depends(get_db)) -> Projetos:
+def post_projetos(projetos: ProjetosRequest, db: Session = Depends(get_db),current_user: User = Depends(get_current_user)) -> Projetos:
     new_projetos = Projetos(
         **projetos.model_dump()
     )
@@ -30,7 +30,11 @@ def post_projetos(projetos: ProjetosRequest, db: Session = Depends(get_db)) -> P
     return new_projetos
 
 @router.put("/{id}", response_model=ProjetosResponse)
-def projetos_update(id:int, projetos_request: ProjetosRequest, db: Session = Depends(get_db)) -> Projetos:
+def projetos_update(id:int, projetos_request: ProjetosRequest, db: Session = Depends(get_db),current_user: User = Depends(get_current_user)) -> Projetos:
+    if current_user.id != id:
+        raise HTTPException(status_code=403, detail="Usuário não autorizado")
+    
+    
     projetos = db.query(Projetos).filter(Projetos.id == id).first()
     if not projetos:
         raise HTTPException(status_code=404, detail="Projeto não encontrado")
@@ -43,14 +47,21 @@ def projetos_update(id:int, projetos_request: ProjetosRequest, db: Session = Dep
 
 
 @router.delete("/{id}", status_code=204)
-def projetos_delete(id:int, db: Session = Depends(get_db)):
+def projetos_delete(id:int, db: Session = Depends(get_db),current_user: User = Depends(get_current_user)):
+    if current_user.id != id:
+        raise HTTPException(status_code=403, detail="Usuário não autorizado")
+    
     projetos = db.query(Projetos).filter(Projetos.id == id).first()
     db.delete(projetos)
     db.commit()
     return projetos
 
 @router.get("/{titulo}", response_model=ProjetosResponse)
-def get_projetos_by_titulo(titulo:str, db:Session = Depends(get_db)) -> ProjetosResponse:
+def get_projetos_by_titulo(titulo:str, db:Session = Depends(get_db),current_user: User = Depends(get_current_user)) -> ProjetosResponse:
+    if current_user.id != id:
+        raise HTTPException(status_code=403, detail="Usuário não autorizado")
+        
+    
     projetos = db.query(Projetos).filter(Projetos.titulo == titulo).first()
     if not projetos:
         raise HTTPException(status_code=404, detail="Projeto não encontrado")
